pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "praevisio"
version = "0.1.6"
description = "A CLI tool for AI governance through verifiable promises."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Praevisio Maintainers" }
]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: Other/Proprietary License",
  "Operating System :: OS Independent",
]
dependencies = [
  "typer>=0.12",
  "PyYAML>=6.0",
  "pytest>=7.4",
  "semgrep",
  "abductio-core>=0.2.3",
]

[project.urls]
Homepage = "https://praevisio.promise.foundation"
Repository = "https://github.com/Promise-Foundation/praevisio"

[tool.hatch.build]
packages = ["src/praevisio"]

# Optional dependency groups for uv
[dependency-groups]
docs = [
  "Sphinx>=7.0",
  "myst-parser>=2.0",
  "sphinx-rtd-theme>=2.0",
  "linkify-it-py>=2.0",
]
dev = [
  "behave>=1.2.6",
]

[project.scripts]
praevisio = "praevisio.presentation.cli:main"

[tool.pytest.ini_options]
norecursedirs = ["fixtures"]
